Move Block Streaming API and handlers
